+++ packages/tx/src/index.ts
@@ -1,4 +1,6 @@
-export { UnsignedLegacyTransaction, SignedLegacyTransaction } from './legacyTransaction'
-export { UnsignedEIP2930Transaction, SignedEIP2930Transaction } from './eip2930Transaction'
+export { default as UnsignedLegacyTransaction } from './unsignedLegacyTransaction'
+export { default as UnsignedEIP2930Transaction } from './unsignedEIP2930Transaction'
+export { default as SignedLegacyTransaction } from './signedLegacyTransaction'
+export { default as SignedEIP2930Transaction } from './signedEIP2930Transaction'
 export { default as TransactionFactory } from './transactionFactory'
 export * from './types'
