+++ packages/tx/src/transactionFactory.ts
@@ -1,6 +1,8 @@
 import Common from '@ethereumjs/common'
-import { SignedLegacyTransaction, UnsignedLegacyTransaction } from './legacyTransaction'
-import { SignedEIP2930Transaction, UnsignedEIP2930Transaction } from './eip2930Transaction'
+import { default as UnsignedLegacyTransaction } from './unsignedLegacyTransaction'
+import { default as UnsignedEIP2930Transaction } from './unsignedEIP2930Transaction'
+import { default as SignedLegacyTransaction } from './signedLegacyTransaction'
+import { default as SignedEIP2930Transaction } from './signedEIP2930Transaction'
 import { TxOptions, Transaction } from './types'
 
 const DEFAULT_COMMON = new Common({ chain: 'mainnet' })
